+++ CHANGES.md
@@ -4,7 +4,7 @@ Changelog
 1.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- add property on wildcard.hps.opensearch.WildcardHPSCatalog for the instance prefix
 
 
 1.1.0 (2022-05-12)